Added ViewModels folder and moved stepvm file there
