+++ ProjectManagement/ProjectManagement.Core/UseCases/Steps/Queries/GetSteps/GetStepsQuery.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using MediatR;
+using ProjectManagement.Core.UseCases.Steps.ViewModels;
 
 namespace ProjectManagement.Core.UseCases.Steps.Queries.GetSteps
 {
